Removed all edge traits except core::Edge.
The various edge traits never ended up making a good API
instead making it difficult to manage the weights of edges.
Therefore, we have removed the use of the edge traits from
other method/function signatures.
The remaining trait (core::Edge) is now simply there to provide
usefull methods for working with a pair of vertices.